+++ docs/changelog.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Upcoming
 
+### Added
+
+- New `six-paginator` component (experimental).
+
 ## 5.3.1 - 2026-03-06
 
 ### Added